requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
pypdf>=3.17.0
chromadb>=0.4.0
python-dotenv>=1.0.0
